BioDemultiplexer

Documentation for BioDemultiplexer.

BioDemultiplexer.classify_sequences

BioDemultiplexer.divide_fastq

BioDemultiplexer.divide_fastq

BioDemultiplexer.execute_demultiplexing

BioDemultiplexer.find_best_matching_bc

BioDemultiplexer.preprocess_bc_file

BioDemultiplexer.classify_sequences — Function

Compare each sequence in the fastqR1 file with the sequences in bcdf, and classify the sequences of the specified file based on that comparison.

BioDemultiplexer.divide_fastq — Method

Divides a single FASTQ file for parallel processing.

BioDemultiplexer.divide_fastq — Method

Divides a pair of FASTQ files into smaller parts for parallel processing. It calculates the number of reads per worker and uses the split command to divide the files.

BioDemultiplexer.execute_demultiplexing — Method

Orchestrates the entire demultiplexing process for FASTQ files. Handles the preprocessing, dividing, demultiplexing, and merging of files.

BioDemultiplexer.find_best_matching_bc — Method

#変えた。 Calculate and compare the similarity of a given sequence seq with the sequences in the given DataFrame bc_df.

Returns

A tuple (max_score_bc, delta), where max_score_bc is the index of the best matching sequence in bc_df, and delta is the difference between the highest and second-highest scores.

BioDemultiplexer.preprocess_bc_file — Function

Preprocesses the barcode file by modifying sequences based on specific criteria.